[mocker_db] minor fixes for initial llm filter; separate async search_database_async method
